dev-work/horse-race-plotter/requirements.txt
matplotlib>=3.5.0
pyyaml>=6.0
